Tighten the animation pictures generation
Add various changes to several files:

-bar2donut script: Now takes a destination folder and an integer id. A
single run of this script will create a set of svgs for a single
animation. The svgs will be written inside the folder specified in
argument (via command line) and will have the integer id embedded in
their name. Also changed the animation duration timer to output a
hundred-ish pictures

-svgtojpg.sh bash script: Now checks for the OS executing the command
and call the appropriate mogrify command

-gen_animations... bash script: This script takes in argument a
destination folder and nb_loops an integer. Calls a for loop with
exactly nb_loops iteration. For each iteration calls the bar2donut
script and the conversion script from svg to jpg (else the number of
elements is too important for the command to work)