Configure database for the transition module
Set up a new table in the database and files/images service.

Refactors also the saving location of images: now the transition images
are all saved under the `transition_files` folder in the static repo,
and split into `input_sketches` and `output_gifs`.

TODO: Update the documentation with the new saving location

Configure placeholders to display previously computed transitions with
the token in URL as arguments (allows communication to peers)